+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 CHANGELOG
 ==========
 
+- 2026-04-13 RestPanel: Fix REST API error display and styling
 - 2026-04-13 RestPanel: Add device power control action panel